+++ tests/Functional/Add2ArchiveTest.php
@@ -8,7 +8,7 @@ class Add2ArchiveTest extends FunctionalTestCase
 {
     private static $file;
 
-    public static function tearDownAfterClass()
+    public static function tearDownAfterClass(): void
     {
         parent::tearDownAfterClass();