+++ tests/Tests/FileStrategy/AbstractFileStrategyTest.php
@@ -8,11 +8,10 @@ use Alchemy\Zippy\Exception\RuntimeException;
 
 class AbstractFileStrategyTest extends TestCase
 {
-    /**
-     * @expectedException   \InvalidArgumentException
-     */
     public function testGetAdaptersWithNoDefinedServices()
     {
+        $this->expectException(\InvalidArgumentException::class);
+
         $container = AdapterContainer::load();
 
         $stub = $this->getMockForAbstractClass('Alchemy\Zippy\FileStrategy\AbstractFileStrategy', array($container));
@@ -24,7 +23,7 @@ class AbstractFileStrategyTest extends TestCase
 
 
         $adapters = $stub->getAdapters();
-        $this->assertInternalType('array', $adapters);
+        $this->assertIsArray($adapters);
         $this->assertCount(0, $adapters);
     }
 
@@ -41,7 +40,7 @@ class AbstractFileStrategyTest extends TestCase
             )));
 
         $adapters = $stub->getAdapters();
-        $this->assertInternalType('array', $adapters);
+        $this->assertIsArray($adapters);
         $this->assertCount(2, $adapters);
         foreach ($adapters as $adapter) {
             $this->assertInstanceOf('Alchemy\\Zippy\\Adapter\\AdapterInterface', $adapter);
@@ -73,7 +72,7 @@ class AbstractFileStrategyTest extends TestCase
             )));
 
         $adapters = $stub->getAdapters();
-        $this->assertInternalType('array', $adapters);
+        $this->assertIsArray($adapters);
         $this->assertCount(1, $adapters);
         foreach ($adapters as $adapter) {
             $this->assertSame($adapterMock, $adapter);