+++ tests/Tests/Resource/ResourceTest.php
@@ -32,7 +32,7 @@ class ResourceTest extends TestCase
         $resource = new Resource($original, $target);
         $result = $resource->canBeProcessedInPlace($context);
 
-        $this->assertInternalType('boolean', $result);
+        $this->assertIsBool($result);
         $this->assertEquals($expected, $result);
     }